services/metadata-extractor/requirements.txt
pika==1.3.2
neo4j==5.14.1
Pillow==10.1.0
exifread==3.0.0
pytest==7.4.3
